begin work to use models.py
